Una encimera se vende por metros, y el modelo pedia un numero entero
Al guardar el presupuesto EXP-2026-158 (dos lineas de ENCIMERA) salia
"[object Object]" y nada mas. Dos fallos encadenados, y el segundo escondia al
primero.

EL DE VERDAD: `quantity` estaba declarado como ENTERO. Una encimera se vende por
METROS LINEALES —3,5 ml es lo normal— y Pydantic rechazaba el guardado entero.

Redondear habria sido PEOR que el fallo: cobrar 4 metros de encimera cuando se
han pedido 3,5 es cambiarle el pedido al cliente sin decirselo, y encima sin que
salte nada. Ahora las lineas de venta admiten decimales.

El DESPIECE sigue siendo entero a proposito: alli una pieza es una pieza, y
media pieza no existe. Va escrito y con su candado.

EL QUE LO ESCONDIA: cuando FastAPI rechaza los datos no contesta con una frase,
contesta con una LISTA de objetos, uno por campo mal. La pantalla hacia
`new Error(d.detail)` y JavaScript convierte un objeto en "[object Object]". O
sea que el servidor decia exactamente que linea y que campo estaban mal, y esa
informacion se tiraba por el camino.

Ahora hay un traductor (services/errores.js) que lo lee y dice "linea 1 ·
cantidad: ...", contando desde 1 porque en la pantalla la primera linea es la 1.
Y un candado barre TODO el frontend para que nadie vuelva a meter un `detail`
crudo en un Error.

783 candados verdes, build con CI=true y cabeceras comprobadas.

+++ backend/models/schemas.py
@@ -341,7 +341,10 @@ class BudgetItemModel(BaseModel):
     productId: str
     productCode: str
     productName: str
-    quantity: int = 1
+    # Decimal por lo mismo que en el digitalizador: hay líneas que se venden por
+    # metros (encimera, zócalo, cantos) y 3,5 es una cantidad legítima. Ojo: en
+    # el DESPIECE sí es entero, porque allí una pieza es una pieza.
+    quantity: float = 1
     customWidth: Optional[float] = None
     customHeight: Optional[float] = None
     customDepth: Optional[float] = None
@@ -880,7 +883,14 @@ class DigitalizadorMatchedProduct(BaseModel):
 class DigitalizadorLine(BaseModel):
     """A single line extracted from the draft"""
     id: str
-    quantity: int = 1
+    # DECIMAL, no entero. Una ENCIMERA se vende por METROS LINEALES: 3,5 ml es
+    # una cantidad de lo más normal. Con `int`, Pydantic rechazaba el guardado
+    # entero y lo que llegaba a la pantalla era «[object Object]», así que ni
+    # siquiera se sabía por qué.
+    #
+    # Redondearla habría sido peor que el fallo: cobrar 4 metros de encimera
+    # cuando se han pedido 3,5 es cambiarle el pedido al cliente sin decírselo.
+    quantity: float = 1
     reference: str = ""
     description: str
     price: float = 0
